commit update admin side.

+++ app/routes/admin.py
@@ -1,3 +1,10 @@
+'''
+Docstring for app.routes.admin
+
+All the admin acces route are in this file
+'''
+
+
 from flask import Blueprint,render_template,session,redirect,flash,url_for,request
 from flask import abort
 from flask_login import current_user
@@ -23,7 +30,7 @@ def admin_required(f):
 
 admin_bp=Blueprint("admin",__name__,url_prefix='/admin')
 
-
+@admin_required
 @admin_bp.route("/dashboard")
 def dashboard():
     user = get_current_user()
@@ -31,34 +38,49 @@ def dashboard():
 
     return render_template("admin/admin_dashboard.html",curr_user=user,blogs=blogs)
 
-@admin_bp.route("/all_blog")
-def allblog():
-    user = get_current_user()
-    blogs= Blog.query.order_by(Blog.id.desc()).all()
-    return render_template("admin/all_blog.html",curr_user=user,blogs=blogs)
+@admin_bp.route("/blogs", methods=["GET", "POST"])
+@admin_required
+def blogs():
+    curr_user = get_current_user()
 
-@admin_bp.route("/all_users")
-def allusers():
-    user = get_current_user()
-    all=User.query.all()
-    return render_template("admin/all_users.html",curr_user=user,all_user=all)
+    if request.method == "POST":
+        blog_id = request.form.get("delete")
+        if blog_id:
+            blog = Blog.query.get_or_404(blog_id)
+            db.session.delete(blog)
+            db.session.commit()
+            flash("Blog deleted")
 
-@admin_bp.route("/deleteuserblog",methods=['POST','GET'])
-def deleteuserblog():
-    curr_user=get_current_user()
-    if request.method=='POST':
-        blog_id=request.form['delete']
-        blog = Blog.query.get_or_404(blog_id)
-        db.session.delete(blog)
-        db.session.commit()
-        return render_template("admin/all_blog.html",curr_user=curr_user)
+    blogs = Blog.query.order_by(Blog.id.desc()).all()
 
-@admin_bp.route("/deleteuser",methods=['POST','GET'])
-def deleteuser():
-    curr_user=get_current_user()
-    if request.method=='POST':
-        user_id=request.form['delete']
-        user=User.query.get_or_404(user_id)
-        db.session.delete(user)
-        db.session.commit()
-        return render_template("admin/all_users.html",curr_user=curr_user)
+    return render_template(
+        "admin/all_blog.html",
+        curr_user=curr_user,
+        blogs=blogs
+    )
+
+    
+@admin_bp.route("/users", methods=["GET", "POST"])
+@admin_required
+def users():
+    curr_user = get_current_user()
+
+    if request.method == "POST":
+        user_id = request.form.get("delete")
+        if user_id:
+            user = User.query.get_or_404(user_id)
+
+            if user.id == curr_user.id:
+                flash("You cannot delete yourself")
+            else:
+                db.session.delete(user)
+                db.session.commit()
+                flash("User deleted")
+
+    users = User.query.all()
+
+    return render_template(
+        "admin/all_users.html",
+        curr_user=curr_user,
+        users=users
+    )
